refactor: (#117) 상점 모달 UI 훅과 컴포넌트 분리
