+++ src/features/shop-purchase/ui/ShopPurchaseModal.tsx
@@ -1,7 +1,7 @@
-import { useState } from 'react';
+// features/shop-purchase/ui/ShopPurchaseModal.tsx
 import { Receipt } from '@/assets';
 import type { Product } from '@/entities/product';
-import { ReceiptCard } from '@/entities/product';
+import { ReceiptItemCarousel } from './ReceiptItemCarousel';
 
 interface ShopPurchaseModalProps {
   onClose: () => void;
@@ -10,25 +10,12 @@ interface ShopPurchaseModalProps {
   userBalance: number;
 }
 
-const ITEMS_PER_PAGE = 3;
-
 export const ShopPurchaseModal = ({
   onClose,
   items,
   totalPrice,
   userBalance,
 }: ShopPurchaseModalProps) => {
-  const [page, setPage] = useState(0);
-  const maxPage = Math.ceil(items.length / ITEMS_PER_PAGE) || 1;
-
-  const currentItems = items.slice(
-    page * ITEMS_PER_PAGE,
-    (page + 1) * ITEMS_PER_PAGE,
-  );
-
-  const handlePrev = () => setPage((p) => Math.max(0, p - 1));
-  const handleNext = () => setPage((p) => Math.min(maxPage - 1, p + 1));
-
   return (
     <div className="fixed inset-0 z-50 flex items-center justify-center font-ssrm font-bold">
       <div
@@ -47,48 +34,7 @@ export const ShopPurchaseModal = ({
 
           <div className="w-[90%] border-b-[3px] border-dashed border-black/80 my-2" />
 
-          <div className="w-full flex flex-col items-center my-2 h-[260px]">
-            <div className="flex items-center justify-between w-full px-4 h-full">
-              <button
-                onClick={handlePrev}
-                disabled={page === 0}
-                className="text-[#cecece] hover:text-gray-400 disabled:opacity-30 text-6xl font-black"
-              >
-                {'<'}
-              </button>
-
-              <div className="flex gap-x-4 w-[512px] justify-center">
-                {currentItems.map((item) => (
-                  <ReceiptCard key={item.id} product={item} />
-                ))}
-                {currentItems.length < ITEMS_PER_PAGE &&
-                  Array.from({
-                    length: ITEMS_PER_PAGE - currentItems.length,
-                  }).map((_, index) => (
-                    <div key={`empty-${index}`} className="w-[160px]" />
-                  ))}
-              </div>
-
-              <button
-                onClick={handleNext}
-                disabled={page === maxPage - 1}
-                className="text-[#cecece] hover:text-gray-400 disabled:opacity-30 text-6xl font-black"
-              >
-                {'>'}
-              </button>
-            </div>
-
-            <div className="flex gap-x-2 mt-1">
-              {Array.from({ length: maxPage }).map((_, index) => (
-                <span
-                  key={index}
-                  className={`px-2 py-0.5 text-sm border border-[#BABABA] ${page === index ? 'border-gray-400 rounded text-gray-500' : 'text-gray-300'}`}
-                >
-                  {index + 1}
-                </span>
-              ))}
-            </div>
-          </div>
+          <ReceiptItemCarousel items={items} />
 
           <div className="w-[90%] border-b-[3px] border-dashed border-black/80 my-4" />
