+++ src/features/shop/index.ts
@@ -2,3 +2,4 @@ export * from './constants/shop.constants';
 export * from './model/types';
 export { useShopFilter } from './model/useShopFilter';
 export { useShopPreview } from './model/useShopPreview';
+export { useProductSorting } from './model/useProductSorting';
